refactored harfile creation
